Change Mileage type to double for precision
Updated the `Mileage` property in the `MobileDepartVehicle` record from `int?` to `double?` to allow for more precise representation of mileage values, accommodating decimal values.

+++ src/Services/Egisz/Frmo/Frmo.Domain/Models/MobileDepartVehicle.cs
@@ -63,7 +63,7 @@ public record MobileDepartVehicle
 
     [XmlElement("mileage")]
     [JsonPropertyName("mileage")]
-    public int? Mileage { get; set; }
+    public double? Mileage { get; set; }
 
     [XmlElement("isNavigation")]
     [JsonPropertyName("isNavigation")]
